+++ plugins/event_source/k8s.py
@@ -604,9 +604,20 @@ class Watcher:
                                         "{}",
                                     )
                                 )
-                                last_applied_configuration = json.loads(
-                                    last_applied_configuration_str
-                                )
+                                # Check if the fields in changed_fields is present and If present, check if last-applied-configuration is present
+                                # parse it as JSON
+                                # If not present, skip the event
+                                if last_applied_configuration_str:
+                                    last_applied_configuration = json.loads(
+                                        last_applied_configuration_str
+                                    )
+                                else:
+                                    self.logger.debug(
+                                        "No last-applied-configuration found, skipping event",
+                                        object_name,
+                                        self.changed_fields,
+                                    )
+                                    continue
                                 if any(
                                     get_nested_value(
                                         last_applied_configuration, field.split(".")
@@ -622,6 +633,7 @@ class Watcher:
                                 else:
                                     self.logger.debug(
                                         "No change detected in fields %s, skipping event",
+                                        object_name,
                                         self.changed_fields,
                                     )
                                     continue
@@ -633,6 +645,7 @@ class Watcher:
                                 self.logger.debug(
                                     "Ignoring MODIFIED event for deleted object %s",
                                     object_name,
+                                    self.changed_fields,
                                 )
                                 continue